fetch function has been renovated

+++ start-nvim.sh
@@ -1,69 +1,48 @@
 #!/bin/bash
 
 function main() {
-    com=${1-:'.'} #Command: all | install | uninstall
-    cur_path=$(realpath $com/neovim)
-    old_path=$(realpath $com/neovim.old)
+    com=${1:-'.'} #Command: path to code location
+    fetch https://github.com/neovim/neovim.git #Main neovim repository
+    exit_ $?
+}
+
+fetch () #params: url [revision]
+{
+    local url=$1
+    [ -z $url ] && echo 'fetch needs URL parameter. Nothing to be done.' >&2 && return 1
+    local rev=$2
     local stat
-    echo Neovim directory: $cur_path
-    echo Neovim backup:    $old_path
-    read -p 'Continue? (y/n) -> ' stat
-    [[ $stat != 'y' && $stat != 'Y' ]] && exit_ 1
-    #[ -d $cur_path ] && cp -r $cur_path $old_path || git clone https://github.com/neovim/neovim.git $cur_path
+    local name=$(basename $url)
+    local repo_name=${name%.git}
+    cur_path=$(realpath $com/$repo_name)
+    old_path=$(realpath $com/${repo_name}.old)
+    echo $repo_name directory: $cur_path
+    echo $repo_name backup:    $old_path
+    read -p "Pulling updates from $url. Continue? (Y/N): " stat && [[ $stat == [yY] || $stat == [yY][eE][sS] ]] || return 1
     if [[ -d $cur_path ]]; then
         pushd $cur_path
         ls -al
-        stat=$(git remote -v 2>/dev/null | grep 'neovim\.git')
+        stat=$(git remote -v 2>/dev/null | grep "$repo_name\.git")
         #stat=$(git status 2>/dev/null)
-        [[ -n $stat ]] && echo $stat
+        [[ -z $stat ]] && echo $cur_path already exist and is not neovim repository, first do \'rm -rf $cur_path\' && return 1
+        git pull
         popd
+    else
+        git clone $url $cur_path
     fi
-    exit 0
-
-    [ -d $old_path ] && rm -rf $old_path
-    cd $cur_path
-    cur_ver=$(git log | head -1 | awk '{print $2}') || exit_ 1
-    fetch $cur_ver
-    case "$com" in
-        all) all
-        ;;
-        install) install
-        ;;
-        uninstall) uninstall
-        ;;
-        *) echo 'Options: all | install | uninstall'
-            exit_ 1
-        ;;
-    esac
-}
-
-all ()
-{
-    uninstall
-    install
-    exit_ 0
-}
-
-fetch () #param: revision
-{
-    local rev=$1
-    git checkout master
-    git pull
+    pushd $cur_path
     git status
-    echo '****************'
-    #git log | egrep 'NVIM v[[:digit:]]?\.[[:digit:]]?\.[[:digit:]]$' | head -n 5 | awk '{print $2}'
-    git checkout $rev
     git branch -a
     git tag --column
     echo '****************'
+    #git log | egrep 'NVIM v[[:digit:]]?\.[[:digit:]]?\.[[:digit:]]$' | head -n 5 | awk '{print $2}'
+    [ -z "$rev" ] && read -p "Choose release: (empty to continue) -> " rev
+    [ -n "$rev" ] && git checkout $rev || return 1
     echo "Your current revision is: $rev"
     echo '****************'
 }
 
-function install() {
-    local rev
-    read -p "Choose release: " rev
-    [ -n "$rev" ] && git checkout $rev || exit_ 1
+function install_nvim() {
     #make CMAKE_BUILD_TYPE=RelWithDebInfo USE_BUNDLED=OFF && \
     make CMAKE_BUILD_TYPE=RelWithDebInfo && \
     sudo make install
@@ -74,13 +53,23 @@ function install() {
     fi
 }
 
-function uninstall() {
-    read -p "Uninstalling your nvim. Continue? (Y/N): " confirm && [[ $confirm == [yY] || $confirm == [yY][eE][sS] ]] || exit_ 1
-    sudo rm  -v /usr/local/bin/nvim
-    sudo rm -rf /usr/local/share/nvim/
+function uninstall_nvim() {
+    local stat
+    read -p "Uninstalling your nvim. Backup here: ${old_path}. Continue? (Y/N): " stat && [[ $stat == [yY] || $stat == [yY][eE][sS] ]] || exit_ 1
+    mkdir -p $cur_path
+    rm -rf $old_path
+    cp -r $cur_path $old_path
+    sudo mv /usr/local/bin/nvim $old_path/bin_nvim
+    sudo mv /usr/local/share/nvim/ $old_path/share_nvim
     sudo rm -rf builds
-    #sudo rm -rf .deps/.ninja_log
     sudo rm -rf .deps
+}
+
+function restore_nvim() {
+    local stat
+    read -p "Restoring your nvim from ${old_path}. Continue? (Y/N): " stat && [[ $stat == [yY] || $stat == [yY][eE][sS] ]] || exit_ 1
+    [ -f $old_path/bin_nvim ] && sudo rm -v /usr/local/bin/nvim && sudo mv $old_path/bin_nvim /usr/local/bin/nvim
+    [ -d $old_path/share_nvim ] && sudo rm -rf /usr/local/share/nvim/ && sudo mv $old_path/share_nvim /usr/local/share/nvim/
 }
 
 function exit_() {
@@ -94,6 +83,13 @@ function exit_() {
     echo "---------------------------------------" |  tee -a $log
     echo "Uninstall: ./start-nvim.sh uninstall" |  tee -a $log
     exit $1
+}
+
+all ()
+{
+    uninstall
+    install
+    exit_ 0
 }
 
 main $@
